Try all kernlog sources when looking for events
We previously switched to only checking the journal log if it
existed but in practice this can contain a limited amount of
information such that we miss events that would otherwise be
detected in kern.log. Therefore if no events are detected in
the journal (-k) we now fallback to trying kern.log as well.